+++ lib/sticky_infinite_list.dart
@@ -1,7 +1,7 @@
 library sticky_infinite_list;
 
 export './widget.dart';
-export './render_new.dart';
+export './render.dart';
 export './models/alignments.dart';
 export './models/sticky_state.dart';
 export './models/types.dart';
